server/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awscli>=1.44.58",
    "bedrock-agentcore>=1.4.6",
    "ddgs>=9.11.4",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "mistral>=21.0.0",
    "mistralai==1.12.4",
    "openapi-spec-validator>=0.8.4",
    "strands-agents>=1.30.0",
    "strands-agents-tools>=0.2.22",
]
